requirements.txt
flask==0.12.2
gevent==1.2.2
Flask-Caching==1.3.3
gevent-websocket==0.10.1
futures==3.1.1
tensorflow==1.10.0
pillow==4.3.0
opencv-python==3.3.0.10
numpy==1.14.1
grpcio==1.14.2
grpcio_tools==1.14.2
requests==2.18.4
pyinstaller==3.3.1
PyCryptodome==3.6.4
pyyaml==3.13